Output rewrite
Change how the network handles output.
Output neurons were previously not a part of the network,
and output was defined by a adjacency matrix where a
single spike in a output-connected network neuron resulted in a spike.

Following this change, output neurons are now part of the network
and have state.